Bugfix on full_envelope<generic_line>
We were using a reference on the env variable, this makes env not
changeable. We needed to change it on our implementation since v[0] on
full_envelope could change.

Solution was to avoid changing v[0] (it's always the same envelope,
though v[0].q can change) and also use a pointer on env, which
facilitates initialization.